removed index and added constants
